backend/requirements.txt
alembic==1.13.3
bcrypt==4.0.1
email-validator==2.2.0
fastapi==0.115.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.3
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
sqlalchemy==2.0.35
uvicorn[standard]==0.30.6
